Reject ORCID accounts without names
The previously existing code fixing #327 did reject ORCID profiles with
private name in the "Researcher" view but was not rejecting such
profiles at sign up. This should now be fixed.

Additionnally, this removes the obsolete "createProfile" page asking for
a name when no name is provided.

Should fix #502 and probably #330.