Switch to pytest-django-4 and use newer pytest
Due to some mixture of pytest and Unittest, we encoutered some obscure errors with newer versions of pytest. With excluding ProtocolTest that does partial testing of protocols this is fixed.

ProtocolTest uses some pytest-fixtures and, when gets invoked multiple
times, does no proper teardownn of the fixture. As a result, the
database is not clean after the ProtocolTest and this can mess up some
of the tests - if they rely on an empty database.

Downside is that we currently have no Zenodo tests, but I think this is
minor important. For one, these tests fail sometimes due to Zenodo
hickups and they invoked some dry-run, so they were not testing the real
thing.