Make shibboleth metadata available to user
Commonly the shibboleth metadata is provided in the session. This is
very fine, but kind of unsuitable for the backend that just expects a
user.
This introduces thus a decorator that copies the shibboleth data from
the session to the user objects of the request. This could also take
place in the backend, but there are rare occasions where we need the
shibboleth metadata.